[add] add controllers and model files for ip and mac addresses
