+++ firewall/manage.py
@@ -7,13 +7,8 @@ from flask_script import Manager
 from flask_swagger_ui import get_swaggerui_blueprint
 
 from app.main import create_app, db, firewall_init
-from app.main.model import tables
-from app.main.model import chains
-from app.main.model import rules
+from app.main.model import mac, ip
 
-from app.main.controllers import tableController as table
-from app.main.controllers import chainsController as chains
-from app.main.controllers import rulesController as rules
 from app.main.controllers import ipController as ip
 from app.main.controllers import macController as mac
 
@@ -22,9 +17,6 @@ app = create_app(os.getenv('BOILERPLATE_ENV') or 'dev')
 SWAGGER_URL = '/api/docs'
 API_URL = '/docs/swagger.yml'
 
-app.register_blueprint(table.bp)
-app.register_blueprint(rules.bp)
-app.register_blueprint(chains.bp)
 app.register_blueprint(ip.bp)
 app.register_blueprint(mac.bp)
